client/requirements.txt
cloudpickle==3.0.0
grpcio==1.60.1
grpcio-tools==1.60.1
protobuf==4.25.3
setuptools==69.1.0
click==8.1.7
